replace ZETIC edge layer with local heuristic anomaly detector
Removes zetic/ entirely and introduces vitals/anomaly.py with the same
scoring logic. Updates imports in vitals/api.py and agent1_monitor.py.
Changes AnomalyEvent.source from zetic_on_device to heuristic.
